debuggging code added

+++ cse333/exercises/exercise1/ex1.c
@@ -45,7 +45,7 @@ double getPi(int n)
     {
         if (i % 2 == 0)
         {
-            result += (double)(4 / (2 * i * (2 * i + 1) * (2 * i + 2)));
+            result += ((double)4) / ((double)(2 * i * (2 * i + 1) * (2 * i + 2)));
         }
         else
         {
